+++ lib/webspicy/specification.rb
@@ -94,8 +94,9 @@ module Webspicy
 end # module Webspicy
 require_relative 'specification/service'
 require_relative 'specification/condition'
-require_relative 'specification/precondition'
-require_relative 'specification/postcondition'
-require_relative 'specification/errcondition'
+require_relative 'specification/pre'
+require_relative 'specification/post'
+require_relative 'specification/err'
+require_relative 'specification/oldies'
 require_relative 'specification/test_case'
 require_relative 'specification/file_upload'