+++ lib/webspicy/specification/oldies.rb
@@ -1,0 +1,4 @@
+require_relative 'oldies/precondition'
+require_relative 'oldies/postcondition'
+require_relative 'oldies/errcondition'
+require_relative 'oldies/bridge'