+++ lib/webspicy/tester/result.rb
@@ -94,13 +94,13 @@ module Webspicy
 
       def check_postconditions!
         service.postconditions.each do |c|
-          check_one! Result::PostconditionMet.new(self, c)
+          check_one! Result::PostconditionMet.new(self, c.bind(tester))
         end
       end
 
       def check_errconditions!
         service.errconditions.each do |c|
-          check_one! Result::ErrconditionMet.new(self, c)
+          check_one! Result::ErrconditionMet.new(self, c.bind(tester))
         end
       end